api: move methods into survey validation service
